改造 Dialog demo
