+++ src/components/DialogDemo.vue
@@ -1,62 +1,21 @@
 <template>
 <div>Dialog 示例</div>
-<h1>示例1</h1>
-<Button @click="toggle">toggle</Button>
-<Dialog v-model:visible="x" :closeOnClickOverlay="false" :ok="f1" :cancel="f2">
-  <template v-slot:content>
-    <strong>hi</strong>
-    <div>hi2</div>
-  </template>
-  <template v-slot:title>
-    <strong>加粗的标题</strong>
-  </template>
-</Dialog>
-<h1>示例2</h1>
-<Button @click="showDialog">show</Button>
+<Demo :component="Dialog1Demo" />
+<Demo :component="Dialog2Demo" />
 </template>
 
 <script lang="ts">
-import Dialog from '../lib/Dialog.vue'
-import Button from '../lib/Button.vue'
-import {
-  ref,
-  h
-} from 'vue'
-import {
-  openDialog
-} from '../lib/openDialog'
+import Demo from './Demo.vue';
+import Dialog1Demo from './Dialog1.demo.vue';
+import Dialog2Demo from './Dialog2.demo.vue';
 export default {
   components: {
-    Dialog,
-    Button
+    Demo
   },
   setup() {
-    const x = ref(false)
-    const toggle = () => {
-      x.value = !x.value
-    }
-    const f1 = () => {
-      return false
-    }
-    const f2 = () => {}
-    const showDialog = () => {
-      openDialog({
-        title: h('strong', {}, '标题'),
-        content: '你好',
-        ok() {
-          console.log('ok')
-        },
-        cancel() {
-          console.log('cancel')
-        }
-      })
-    }
     return {
-      x,
-      toggle,
-      f1,
-      f2,
-      showDialog
+      Dialog1Demo,
+      Dialog2Demo
     }
   }
 }
